feat: keep failed jobs if in developer mode

+++ Model/Cron.php
@@ -3,6 +3,7 @@
 
 namespace Discorgento\Queue\Model;
 
+use Discorgento\Core\Helper\Data as CoreHelper;
 use Discorgento\Queue\Contracts\JobInterface;
 use Discorgento\Queue\Model\ResourceModel\Message\CollectionFactory as MessageCollectionFactory;
 use Magento\Framework\ObjectManagerInterface;
@@ -10,6 +11,9 @@ use Psr\Log\LoggerInterface;
 
 class Cron
 {
+    /** @var CoreHelper */
+    protected $coreHelper;
+
     /** @var LoggerInterface */
     protected $logger;
 
@@ -23,11 +27,13 @@ class Cron
     protected $objectManager;
 
     public function __construct(
+        CoreHelper $coreHelper,
         LoggerInterface $logger,
         MessageCollectionFactory $messageCollectionFactory,
         MessageRepository $messageRepository,
         ObjectManagerInterface $objectManager
     ) {
+        $this->coreHelper = $coreHelper;
         $this->logger = $logger;
         $this->messageCollectionFactory = $messageCollectionFactory;
         $this->messageRepository = $messageRepository;
@@ -43,6 +49,11 @@ class Cron
                 /** @var JobInterface */
                 $job = $this->objectManager->create($message->getJobClass());
                 $job->execute($message->getTarget(), $message->getAdditionalData());
+
+                // if in developer mode only clean jobs on success
+                if ($this->coreHelper->isDeveloperMode()) {
+                    $this->messageRepository->delete($message);
+                }
             } catch (\Throwable $th) {
                 $errorMessage = "Job {$message->getJobClass()} failed: '{$th->getMessage()}'";
                 $this->logger->error($errorMessage, [
@@ -51,7 +62,10 @@ class Cron
                 ]);
             }
 
-            $this->messageRepository->delete($message);
+            // keep failed jobs when in developer mode
+            if ($this->coreHelper->isProductionMode()) {
+                $this->messageRepository->delete($message);
+            }
         }
 
         return $this;
